printf: read %z and %t as size_t/ptrdiff_t, not as long
src/stdio/printf.c read both length modifiers with va_arg(ap, long) at
three sites.  On this target long is 32 bits and size_t/ptrdiff_t are
64 (LLP64), so the type was simply wrong.

fprintf.html, the length modifiers: "z  Specifies that a following d,
i, o, u, x, or X conversion specifier applies to a size_t or the
corresponding signed integer type argument; or that a following n
conversion specifier applies to a pointer to a signed integer type
corresponding to a size_t argument."  t says the same for ptrdiff_t.

The %n case is the worst of the three and is not a formatting defect at
all: it stored through *(long *), writing FOUR bytes into the caller's
EIGHT-byte size_t and leaving the upper four whatever they were --
silent corruption of an object the caller owns.

The tree already contained the correct pattern in the file implementing
the same grammar: src/stdio/scanf.c pairs LM_z with size_t and LM_t
with ptrdiff_t and has no instance of this.  printf.c was the only
offender.  Fixed by pulling each as the type the page names, and
splitting the shared `case LM_z: case LM_t:` arms since the two are no
longer the same type.

THE FENCE'S TEST ASSERTED A WRONG CONSTANT.  It checked "%zu" of
(size_t)0xdeadbeef << 32 against the literal "16045690981097537536".
That value is 0xdeadbeef00020000; the expression is 0xdeadbeef00000000
= 16045690981097406464.  So a CORRECT fix still failed the test --
which invites exactly the wrong move, of adjusting the library until
the bad literal comes out.  Caught by noticing the fixed %zu agreed
with %llu of the same value, then checking the arithmetic
independently.

The test now checks two ways: against "%llu" of the same value, which
cannot be typo'd out of agreement, and against the corrected literal,
which would still catch a bug shared by both conversions.

Extended, because the fence's version could not distinguish a correct
fix from the defect on the signed path: it used only -1, which
sign-extends to the same bits through a 32-bit long or a 64-bit
ptrdiff_t.  Mutation-testing showed a -1-only check passing with the
signed defect fully restored.  It now uses 0xdeadbeef, which a 32-bit
long reads back as negative, and keeps a -1 case so the fix cannot
degenerate into "always widen wrongly".  %td and %tn are covered
alongside %zd and %zn; the fence tested neither t form.

A SECOND TEST WAS RELYING ON THE DEFECT.  test/stdio.c:72 passed
`(long)-7` to "%td" -- a type mismatch, and undefined behaviour, which
worked only while printf read %t as long.  With the fix it printed
4294967289: the 32-bit -7 read as a 64-bit ptrdiff_t.  Every
neighbouring line in that block already passes the type its conversion
names (%zu gets size_t, %zd gets ssize_t, %jd gets long long); this one
alone did not.  Corrected to (ptrdiff_t)-7, with the reason recorded so
it is not "simplified" back.

Reproduced: 2 failing checks before, all pass after.  Mutation-tested,
three, one per site, all caught: unsigned path back to long (2
failures), %n back to *(long *) (2), signed path back to long (2).

make check: 58 passed, 0 failed, 4 unverified.

+++ test/stdio.c
@@ -69,7 +69,13 @@ static void test_printf(void)
 	FMT("-1", "%hd", 65535);
 	FMT("65535", "%hu", 65535);
 	FMT("99", "%jd", (long long)99);
-	FMT("-7", "%td", (long)-7);
+	/* (ptrdiff_t), not (long): %td requires a ptrdiff_t argument, and on
+	 * LLP64 long is 32 bits while ptrdiff_t is 64, so passing a long was
+	 * a type mismatch -- undefined behaviour that happened to work only
+	 * while printf.c read %t as long (fixed in "printf: read %z and %t
+	 * as size_t/ptrdiff_t, not as long").  Every neighbouring line here
+	 * already passes the type its conversion names; this one did not. */
+	FMT("-7", "%td", (ptrdiff_t)-7);
 
 	/* flags, width, precision */
 	FMT("   42", "%5d", 42);
